TresureHunt/requirements.txt
fastapi
uvicorn[standard]
slowapi
httpx
requests
python-dotenv
aiofiles
